# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GuardRail API",
    "description": "Agent 写操作安全执行层:模型只提议,系统做决策。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/runs": {
      "post": {
        "tags": [
          "governance"
        ],
        "summary": "登记一次执行(此时无任何副作用)",
        "operationId": "create_api_runs_post",
        "parameters": [
          {
            "name": "x-actor",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "governance"
        ],
        "summary": "执行记录列表",
        "operationId": "list_runs_api_runs_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RunStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_uid}": {
      "get": {
        "tags": [
          "governance"
        ],
        "summary": "执行详情(含步骤与断点)",
        "operationId": "get_run_api_runs__run_uid__get",
        "parameters": [
          {
            "name": "run_uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Uid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_uid}/execute": {
      "post": {
        "tags": [
          "governance"
        ],
        "summary": "内联推进(与 worker 同一套执行器)",
        "operationId": "execute_run_api_runs__run_uid__execute_post",
        "parameters": [
          {
            "name": "run_uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Uid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecuteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_uid}/steps/{seq}/approve": {
      "post": {
        "tags": [
          "governance"
        ],
        "summary": "批准挂起的步骤",
        "operationId": "approve_step_api_runs__run_uid__steps__seq__approve_post",
        "parameters": [
          {
            "name": "run_uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Uid"
            }
          },
          {
            "name": "seq",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Seq"
            }
          },
          {
            "name": "x-actor",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_uid}/steps/{seq}/retry": {
      "post": {
        "tags": [
          "governance"
        ],
        "summary": "重跑单步(幂等兜底)",
        "operationId": "retry_step_api_runs__run_uid__steps__seq__retry_post",
        "parameters": [
          {
            "name": "run_uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Uid"
            }
          },
          {
            "name": "seq",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Seq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetryStepResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateRunRequest": {
        "properties": {
          "goal": {
            "type": "string",
            "maxLength": 500,
            "minLength": 2,
            "title": "Goal"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/PlannedStepIn"
            },
            "type": "array",
            "title": "Steps"
          },
          "from_intent": {
            "type": "boolean",
            "title": "From Intent",
            "description": "忽略 steps,由确定性规划器从 goal 生成一步计划(W3 才换成 LLM)",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "goal"
        ],
        "title": "CreateRunRequest"
      },
...
      "ExecuteResponse": {
        "properties": {
          "run": {
            "$ref": "#/components/schemas/RunDetail"
          },
          "resumed_from": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resumed From"
          },
          "executed": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Executed"
          },
          "replayed": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Replayed"
          }
        },
        "type": "object",
        "required": [
          "run"
        ],
        "title": "ExecuteResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlannedStepIn": {
        "properties": {
          "seq": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Seq"
          },
          "tool": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Tool"
          },
          "args": {
            "additionalProperties": true,
            "type": "object",
            "title": "Args"
          },
          "requires_approval": {
            "type": "boolean",
            "title": "Requires Approval",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "seq",
          "tool"
        ],
        "title": "PlannedStepIn"
      },
...
      "RetryStepResponse": {
        "properties": {
          "run": {
            "$ref": "#/components/schemas/RunDetail"
          },
          "seq": {
            "type": "integer",
            "title": "Seq"
          },
          "replayed": {
            "type": "boolean",
            "title": "Replayed"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "run",
          "seq",
          "replayed"
        ],
        "title": "RetryStepResponse"
      },
      "RunDetail": {
        "properties": {
          "run_uid": {
            "type": "string",
            "title": "Run Uid"
          },
          "goal": {
            "type": "string",
            "title": "Goal"
          },
          "actor": {
            "type": "string",
            "title": "Actor"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "plan": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Plan"
          },
          "checkpoint_seq": {
            "type": "integer",
            "title": "Checkpoint Seq"
          },
          "checkpoint": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkpoint"
          },
          "waiting_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Waiting Ref"
          },
          "approved_seqs": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Approved Seqs"
          },
          "attempt": {
            "type": "integer",
            "title": "Attempt"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/StepView"
            },
            "type": "array",
            "title": "Steps"
          }
        },
        "type": "object",
        "required": [
          "run_uid",
          "goal",
          "actor",
          "trace_id",
          "status",
          "plan",
          "checkpoint_seq",
          "approved_seqs",
          "attempt",
          "created_at",
          "steps"
        ],
        "title": "RunDetail"
      },
      "RunListResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/RunView"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "limit",
          "offset",
          "items"
        ],
        "title": "RunListResponse"
      },
      "RunStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "RUNNING",
          "WAITING_APPROVAL",
          "SUCCEEDED",
          "FAILED",
          "COMPENSATED"
        ],
        "title": "RunStatus",
        "description": "一次 Agent 执行的生命周期。\n\n这是「可写执行层」的状态中枢:所有崩溃恢复都从这几个状态推导,\n而不是靠进程内的内存标记。"
      },
      "RunView": {
        "properties": {
          "run_uid": {
            "type": "string",
            "title": "Run Uid"
          },
          "goal": {
            "type": "string",
            "title": "Goal"
          },
          "actor": {
            "type": "string",
            "title": "Actor"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "plan": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Plan"
          },
          "checkpoint_seq": {
            "type": "integer",
            "title": "Checkpoint Seq"
          },
          "checkpoint": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkpoint"
          },
          "waiting_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Waiting Ref"
          },
          "approved_seqs": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Approved Seqs"
          },
          "attempt": {
            "type": "integer",
            "title": "Attempt"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "run_uid",
          "goal",
          "actor",
          "trace_id",
          "status",
          "plan",
          "checkpoint_seq",
          "approved_seqs",
          "attempt",
          "created_at"
        ],
        "title": "RunView"
      },
      "StepKind": {
        "type": "string",
        "enum": [
          "EXECUTE",
          "APPROVAL"
        ],
        "title": "StepKind"
      },
      "StepStatus": {
        "type": "string",
        "enum": [
          "RUNNING",
          "SUCCEEDED",
          "FAILED"
        ],
        "title": "StepStatus"
      },
      "StepView": {
        "properties": {
          "seq": {
            "type": "integer",
            "title": "Seq"
          },
          "kind": {
            "$ref": "#/components/schemas/StepKind"
          },
          "status": {
            "$ref": "#/components/schemas/StepStatus"
          },
          "tool_name": {
            "type": "string",
            "title": "Tool Name"
          },
          "args": {
            "additionalProperties": true,
            "type": "object",
            "title": "Args"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "attempt": {
            "type": "integer",
            "title": "Attempt"
          },
          "replayed": {
            "type": "boolean",
            "title": "Replayed"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "seq",
          "kind",
          "status",
          "tool_name",
          "args",
          "attempt",
          "replayed"
        ],
        "title": "StepView"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}